validate default value
